requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
asteroid-filterbanks==0.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
backcall==0.2.0
backports.cached-property==1.0.2
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorlog==6.8.2
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.7
decorator==5.1.1
dill==0.3.8
distro==1.9.0
docopt==0.6.2
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
greenlet==3.1.1
grpcio==1.67.0
h11==0.14.0
hmmlearn==0.2.8
httpcore==1.0.6
httpx==0.27.2
huggingface==0.0.1
huggingface-hub==0.26.2
HyperPyYAML==1.2.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==7.34.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
julius==0.2.7
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.9.2
lightning==2.3.3
lightning-fabric==2.4.0
lightning-utilities==0.11.7
llvmlite==0.41.1
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==2.8.8
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.54.3
optuna==4.0.0
packaging==24.1
pandas==2.0.3
parso==0.8.4
peft==0.13.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==23.0.1
platformdirs==4.3.6
pooch==1.8.2
primePy==1.3
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==3.20.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyDeprecate==0.3.2
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-lightning==1.6.5
pytorch-metric-learning==2.7.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
resampy==0.4.3
rich==13.9.2
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
semver==3.0.2
sentencepiece==0.2.0
setuptools==56.0.0
shellingham==1.5.4
simplejson==3.19.3
singledispatchmethod==1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==0.5.16
SQLAlchemy==2.0.35
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.4.1
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.4.1
torchmetrics==0.11.4
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.2
triton==3.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.44.0
whisper==1.1.10
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
